requirements.txt
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
future==0.17.1
gym==0.10.8
idna==2.8
numpy==1.16.4
numpy-stl==2.10.1
protobuf==3.8.0
pyglet==1.3.2
python-utils==2.3.0
requests==2.22.0
six==1.12.0
tensorboardX==1.7
urllib3==1.25.3
matplotlib
gitpython
torchvision
seaborn
tqdm